utils/file_utils.py
import os
import json
import yaml

def ensure_dir_exists(dir_path):
    """
    Ensures that a directory exists, creating it if necessary.
    """
    if not os.path.exists(dir_path):
        os.makedirs(dir_path)
        print(f"Created directory: {dir_path}")

def read_json(file_path):
    """Reads a JSON file."""
    try:
        with open(file_path, 'r', encoding='utf-8') as f:
            return json.load(f)
    except FileNotFoundError:
        print(f"Error: File not found at {file_path}")
        return None
    except json.JSONDecodeError:
        print(f"Error: Could not decode JSON from {file_path}")
        return None

def write_json(data, file_path):
    """Writes data to a JSON file."""
    ensure_dir_exists(os.path.dirname(file_path))
    with open(file_path, 'w', encoding='utf-8') as f:
        json.dump(data, f, ensure_ascii=False, indent=4)
    print(f"Data written to {file_path}")

def read_yaml(file_path):
    """Reads a YAML file."""
    try:
        with open(file_path, 'r', encoding='utf-8') as f:
            return yaml.safe_load(f)
    except FileNotFoundError:
        print(f"Error: File not found at {file_path}")
        return None
    except yaml.YAMLError as e:
        print(f"Error: Could not parse YAML from {file_path}: {e}")
        return None

# Add other file utility functions as needed
